Restructure some of the exploit methods.
