+++ lib/msf/core/exploit/mixins.rb
@@ -98,7 +98,8 @@ require 'msf/core/exploit/winrm'
 # WebApp
 require 'msf/core/exploit/web'
 
-# Firefox addons
+# Firefox
+require 'msf/core/exploit/remote/firefox_privilege_escalation'
 require 'msf/core/exploit/remote/firefox_addon_generator'
 
 require 'msf/core/exploit/remote/browser_exploit_server'
